Another clean up
