+++ UnityProject/Assets/Scripts/Electricity/ElectricalManager.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class ElectricalManager : MonoBehaviour {
-	void Update () {
+	void Update () { //since you can't add MonoBehaviour Too static
 		ElectricalSynchronisation.Update ();
 	}
 }
