+++ UnityProject/Assets/Scripts/Electricity/FunctionsAndClasses.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 25eccd63b3c5c0a419ef90c96f3565ee
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
