+++ UnityProject/Assets/Scripts/Electricity/FunctionsAndClasses/BatteryCalculation.cs
