+++ UnityProject/Assets/Scripts/Electricity/FunctionsAndClasses/BoolClass.cs
