+++ UnityProject/Assets/Scripts/Electricity/FunctionsAndClasses/DeadEndConnection.cs
