+++ UnityProject/Assets/Scripts/Electricity/FunctionsAndClasses/ElectricalDataCleanup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5aa35af6e57e9dc4997e720e2fc74b3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
