+++ UnityProject/Assets/Scripts/Electricity/FunctionsAndClasses/FloatClass.cs
