+++ UnityProject/Assets/Scripts/Electricity/FunctionsAndClasses/IntrinsicElectronicData.cs
