+++ UnityProject/Assets/Scripts/Electricity/FunctionsAndClasses/TransformerCalculations.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3ab0184aeac28e24aba6b96fafbc8b4e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
