+++ UnityProject/Assets/Scripts/Electricity/Interfaces/InputOutputFunctions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 43f7426145ead444abfd544d005aef29
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
